requirements.txt
requests
pandas
numpy
matplotlib
selenium
webdriver_manager
playwright-stealth